Se añade la paginación y el filtro por estado y tipo de prestamo

+++ domain/model/src/main/java/co/com/bancolombia/model/loantype/LoanType.java
@@ -18,7 +18,7 @@ public class LoanType {
     private String name;
     private BigDecimal minAmount;
     private BigDecimal maxAmount;
-    private Double interestRate;
+    private BigDecimal interestRate;
     private Boolean automaticValidation;
 
 }
